Pull out REL
